完成回复功能

+++ src/main/java/com/ljy/community/controller/PublishController.java
@@ -12,10 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-
-
 import javax.servlet.http.HttpServletRequest;
-import javax.sql.rowset.spi.SyncResolver;
 
 @Controller
 public class PublishController {
@@ -25,7 +22,7 @@ public class PublishController {
 
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,
+    public String edit(@PathVariable(name = "id")Long id,
                        Model model){
         QuestionDTO question=questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -47,7 +44,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value ="description",required = false) String description,
             @RequestParam(value ="tag",required = false) String tag,
-            @RequestParam(value = "id",required = false)Integer id,
+            @RequestParam(value = "id",required = false)Long id,
             HttpServletRequest request,
             Model model){
         model.addAttribute("title",title);
